BitFarming Token
BitFarming is a project that aims to look at environmental changes, which will help the whole world in Plant trees and the construction of wells . Protecting the whole world and the future with global warming. Our goal is to be the most used CryptoCurrency in the world of Gaming with Green ideas.. BF is a hyper-deflationary decentralized buyback token on BSC (Binance Smart Chain) BlockChain. BitFarming will be a P2E project which will be the most known and most used cryptocurrency in the Gaming world.